+++ src/aks-preview/azext_aks_preview/custom.py
@@ -2286,6 +2286,7 @@ def aks_agentpool_add(
     disable_windows_outbound_nat=False,
     allowed_host_ports=None,
     asg_ids=None,
+    enable_managed_dranet=False,
     node_public_ip_tags=None,
     enable_artifact_streaming=False,
     enable_managed_gpu=False,
@@ -2370,6 +2371,7 @@ def aks_agentpool_update(
     # extensions
     allowed_host_ports=None,
     asg_ids=None,
+    enable_managed_dranet=False,
     enable_artifact_streaming=False,
     disable_artifact_streaming=False,
     enable_managed_gpu=False,